Fix makemigrations warning for user model case
Fix issue where a warning is raised when running makemigrations
when user model is passed with different case to the default.
This commonly happens when the default is `auth.User` and
`auth.user` is passed in the `to` field.

The fix is to detect the case difference, and update the incoming
args to match the default.

Relates to #43.